implemented designs for page

+++ myProject/my-college-app/src/App.jsx
@@ -13,14 +13,23 @@ function App() {
 
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('')
+  const [confirmPassword, setConfirmPassword] = useState('')
   let navigate = useNavigate();
 
+  useEffect(() => {
+    let authToken = sessionStorage.getItem('Auth Token')
+    if (authToken) {
+      navigate('/college')
+    }
+  }, [])
+
   const handleAction = (id) => {
-    console.log(id)
     const authentication = getAuth();
     // Register
     if (id === 2) {
-      createUserWithEmailAndPassword(authentication, email, password)
+      let confirmPassCheck = !(confirmPassword.replace(/\s{2,}/g,' ').trim()) ? false : true
+      if(password === confirmPassword && validationForm() && confirmPassCheck) {
+        createUserWithEmailAndPassword(authentication, email, password)
         .then((response) => {
           navigate('/college')
           sessionStorage.setItem('Auth Token', response._tokenResponse.refreshToken)
@@ -30,32 +39,42 @@ function App() {
             toast.error('Email Already in Use');
           }
         })
+      } else {
+        toast.error('Mismatch in Password');
+      }
+     
     }
     // Login
     if (id === 1) {
-      signInWithEmailAndPassword(authentication, email, password)
-        .then((response) => {
-          navigate('/college')
-          sessionStorage.setItem('Auth Token', response._tokenResponse.refreshToken)
-        })
-        .catch((error) => {
-          if(error.code === 'auth/wrong-password'){
-            toast.error('Please check the Password');
-          }
-          if(error.code === 'auth/user-not-found'){
-            toast.error('Please check the Email');
-          }
-        })
+      if(validationForm()) {
+        signInWithEmailAndPassword(authentication, email, password)
+          .then((response) => {
+            navigate('/college')
+            sessionStorage.setItem('Auth Token', response._tokenResponse.refreshToken)
+          })
+          .catch((error) => {
+            if(error.code === 'auth/wrong-password'){
+              toast.error('Please check the Password');
+            }
+            if(error.code === 'auth/user-not-found'){
+              toast.error('Please check the Email');
+            }
+          })
+      }
     }
   }
 
-  useEffect(() => {
-    let authToken = sessionStorage.getItem('Auth Token')
-
-    if (authToken) {
-      navigate('/college')
+  function validationForm() {
+    if(!(email.replace(/\s{2,}/g,' ').trim())) {
+      toast.error('Email Can not be empty');
+      return false;
     }
-  }, [navigate])
+    if(!(password.replace(/\s{2,}/g,' ').trim())) {
+      toast.error('Password Can not be empty');
+      return false;
+    }
+    return true;
+  }
 
   return (
     <>
@@ -68,7 +87,8 @@ function App() {
                 title="Login"
                 setEmail={setEmail}
                 setPassword={setPassword}
-                handleAction={() => handleAction(1)} />}
+                handleAction={() => handleAction(1)}
+                isLogin={true} />}
           />
           <Route
             path='/register'
@@ -77,7 +97,9 @@ function App() {
                 title="Register"
                 setEmail={setEmail}
                 setPassword={setPassword}
-                handleAction={() => handleAction(2)} />}
+                setConfirmPassword={setConfirmPassword}
+                handleAction={() => handleAction(2)}
+                isLogin={false} />}
             />
           <Route path='/college' element={<Home />} />
           <Route path='/college/:collegeName' element={<CollegeDetail />} />
